Fixed UI and Card layout, WIP
