+++ go-fish/src/components/Home/Lobby/Game/Game.css
@@ -1,0 +1,335 @@
+/* Base Game Container and Background */
+.game-container {
+  min-height: 100vh;
+  background: radial-gradient(circle at center, #2a2f4e 0%, #161b33 100%);
+  padding: 2rem;
+  color: white;
+  font-family: 'Inter', sans-serif;
+  position: relative;
+  overflow: hidden;
+}
+
+/* Ambient background animation */
+.game-container::before {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  opacity: 0.1;
+  animation: backgroundNoise 8s linear infinite;
+}
+
+.game-board {
+  position: relative;
+  width: 95vw;
+  height: 95vh;
+  max-width: 1600px;
+  margin: 0 auto;
+  background: rgba(255, 255, 255, 0.03);
+  border-radius: 2rem;
+  backdrop-filter: blur(10px);
+  box-shadow: 
+    0 0 100px rgba(0, 150, 255, 0.1),
+    inset 0 0 40px rgba(0, 0, 0, 0.2);
+  overflow: hidden;
+}
+
+/* Turn Order Container */
+.turn-status {
+  position: absolute;
+  top: 2rem;
+  right: 2rem;
+  background: rgba(255, 255, 255, 0.1);
+  backdrop-filter: blur(8px);
+  border-radius: 12px;
+  padding: 1.5rem;
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  min-width: 240px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+.current-turn, .next-turn {
+  margin-bottom: 0.5rem;
+  font-size: 1rem;
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.turn-indicator {
+  width: 8px;
+  height: 8px;
+  border-radius: 50%;
+  background: #4fd1c5;
+  animation: pulseBorder 2s infinite;
+}
+
+/* Players List Container */
+.players-container {
+  position: absolute;
+  left: 2rem;
+  top: 50%;
+  transform: translateY(-50%);
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  max-width: 300px;
+}
+
+.player-card {
+  background: rgba(255, 255, 255, 0.1);
+  backdrop-filter: blur(8px);
+  border-radius: 12px;
+  padding: 1.5rem;
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  transition: all 0.3s ease;
+}
+
+.player-card.current-turn {
+  background: rgba(79, 209, 197, 0.2);
+  border-color: #4fd1c5;
+  box-shadow: 0 0 20px rgba(79, 209, 197, 0.3);
+}
+
+/* Enhanced Card Styling */
+.card {
+  position: relative;
+  width: 120px;
+  height: 168px;
+  background: white;
+  border-radius: 12px;
+  transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
+  transform-origin: center bottom;
+  box-shadow: 
+    0 4px 6px rgba(0, 0, 0, 0.1),
+    0 1px 3px rgba(0, 0, 0, 0.08);
+  will-change: transform;
+  cursor: pointer;
+}
+
+.card svg {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 100%;
+  height: 60%;
+  z-index: 2;
+  color: var(--card-color, #000);
+  filter: drop-shadow(0 1px 1px rgba(0, 0, 0, 0.1));
+}
+
+.card[data-suit="Hearts"] svg,
+.card[data-suit="Diamonds"] svg {
+  --card-color: #dc2626;
+}
+
+.card[data-suit="Spades"] svg,
+.card[data-suit="Clubs"] svg {
+  --card-color: #1a202c;
+}
+
+/* Card Back Pattern */
+.card::before {
+  content: '';
+  position: absolute;
+  inset: 8px;
+  border-radius: 8px;
+  background: 
+    linear-gradient(135deg, #f1f5f9 0%, #e2e8f0 100%),
+    repeating-linear-gradient(
+      45deg,
+      transparent 0,
+      transparent 10px,
+      rgba(0, 0, 0, 0.05) 10px,
+      rgba(0, 0, 0, 0.05) 20px
+    );
+  z-index: 1;
+}
+
+.card.face-up::before {
+  display: none;
+}
+
+/* Player Hand Layout */
+.player-hand {
+  position: absolute;
+  bottom: 2rem;
+  left: 220%;  /* Change from right: 50% to left: 50% */
+  transform: translateX(-50%);  /* Keep this the same */
+  display: flex;
+  gap: 0.5rem;
+  padding: 2rem;
+  min-height: 100px;
+  perspective: 1000px;
+}
+
+.player-hand .card {
+  --card-fan-radius: 400px;
+  --card-fan-angle: 60deg;
+  position: absolute;
+  transform-origin: bottom center;
+}
+
+.player-hand .card:hover {
+  z-index: 40;
+}
+
+.card.selected {
+  transform: translateY(-30px) scale(1.1) !important;
+  box-shadow: 
+    0 20px 25px -5px rgba(0, 0, 0, 0.2),
+    0 8px 10px -6px rgba(0, 0, 0, 0.1),
+    0 0 0 2px #4fd1c5,
+    0 0 20px rgba(79, 209, 197, 0.5);
+  z-index: 50 !important;
+}
+
+/* Game Controls */
+.game-controls {
+  position: fixed;
+  bottom: 2rem;
+  left: 50%;
+  transform: translateX(-50%);
+  display: flex;
+  gap: 1rem;
+  background: rgba(17, 24, 39, 0.95);
+  padding: 1rem 2rem;
+  border-radius: 12px;
+  backdrop-filter: blur(10px);
+  box-shadow: 
+    0 10px 15px -3px rgba(0, 0, 0, 0.1),
+    0 4px 6px -2px rgba(0, 0, 0, 0.05);
+  z-index: 100;
+}
+
+.player-select {
+  padding: 0.75rem 1.5rem;
+  border-radius: 8px;
+  background: rgba(255, 255, 255, 0.1);
+  color: white;
+  border: 1px solid rgba(79, 209, 197, 0.3);
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.ask-button {
+  padding: 0.75rem 2rem;
+  border-radius: 8px;
+  background: linear-gradient(135deg, #4fd1c5, #319795);
+  color: white;
+  font-weight: 600;
+  border: none;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.ask-button:not(:disabled):hover {
+  transform: translateY(-2px);
+  box-shadow: 
+    0 10px 15px -3px rgba(0, 0, 0, 0.1),
+    0 4px 6px -2px rgba(0, 0, 0, 0.05);
+}
+
+/* Deck Counter */
+.deck-counter {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  background: rgba(255, 255, 255, 0.1);
+  backdrop-filter: blur(8px);
+  padding: 1.5rem;
+  border-radius: 12px;
+  text-align: center;
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  min-width: 200px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+/* Animations */
+@keyframes pulseBorder {
+  0%, 100% { box-shadow: 0 0 0 0 rgba(79, 209, 197, 0.4); }
+  50% { box-shadow: 0 0 0 4px rgba(79, 209, 197, 0); }
+}
+
+@keyframes dealCard {
+  0% {
+    opacity: 0;
+    transform: translate(-50%, -50%) scale(0.8) rotate(0deg);
+  }
+  100% {
+    opacity: 1;
+    transform: translate(var(--final-x), var(--final-y)) rotate(var(--final-rotation));
+  }
+}
+
+@keyframes cardFlip {
+  0% { transform: rotateY(0deg); }
+  100% { transform: rotateY(360deg); }
+}
+
+/* Responsive Design */
+@media (max-width: 768px) {
+  .game-board {
+    width: 100vw;
+    height: 100vh;
+    border-radius: 0;
+  }
+
+  .player-hand {
+    transform: scale(0.8);
+  }
+
+  .game-controls {
+    width: 90%;
+    flex-direction: column;
+    gap: 0.75rem;
+  }
+  
+  .turn-status,
+  .players-container {
+    transform: scale(0.9);
+  }
+}
+
+.sets-display {
+  margin-top: 1rem;
+  display: flex;
+  flex-wrap: wrap;
+  gap: 1rem;
+  max-width: 300px;
+}
+
+.set-container {
+  background: rgba(255, 255, 255, 0.1);
+  border-radius: 0.5rem;
+  padding: 0.75rem;
+  backdrop-filter: blur(8px);
+}
+
+.set-header {
+  font-size: 0.875rem;
+  color: rgba(255, 255, 255, 0.9);
+  margin-bottom: 0.5rem;
+}
+
+.set-cards {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 0.5rem;
+}
+
+.set-card {
+  background: rgba(255, 255, 255, 0.05);
+  border-radius: 0.25rem;
+  padding: 0.5rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
